♻️ Refactor store
