+++ src/modules/core/modules/redux/config/index.ts
@@ -1,0 +1,1 @@
+export * from './store'
